requirements.txt
deepgram-sdk
playsound==1.2.2
langchain-groq
langchain_google_genai
langchain_chroma
chromadb
Pillow
pyperclip
pyaudio
python-dotenv
instructor
pydantic
colorama
tavily-python
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
openai==0.28.0
pvporcupine
wave
requests
gtts
gspread